+++ server/src/main/kotlin/com/hockey/elo/elotracker/user/model/UserDTO.kt
@@ -3,7 +3,7 @@ package com.hockey.elo.elotracker.user.model
 data class UserDTO(val user: User) {
     var id: Long = user.id
     var name: String = user.name
-    var elo: Long = user.elo
-    var wins: Long = user.wins
-    var losses: Long = user.losses
+    var elo: Int = user.elo
+    var wins: Int = user.wins
+    var losses: Int = user.losses
 }
